Do k-nearest from https://youtu.be/CI60af3hhS8
A nice tutorial. Learned about some `Vec` methods like `swap`. I also got
to know the borrow checker a bit more with the `k_nearest` function. I
didn't originally pick on the fact that the points data was moved into
the function. So the original points vector was basically changed as it
was iterated through and all the points were stored into the `MaxHeap`.